AU-1771: Unregistered HP profile update callbacks
