+++ public/modules/custom/grants_profile/grants_profile.module
@@ -110,6 +110,7 @@ function grants_profile_theme(): array {
     'variables' => [
       'myProfile' => NULL,
       'editHelsinkiProfileLink' => NULL,
+      'updatelink' => NULL,
     ],
   ];
 
@@ -253,6 +254,7 @@ function grants_profile_preprocess_own_profile_form_unregistered_community(&$var
   $variables['profile'] = $variables["form"]["#profilecontent"];
 
   $variables['basic_info'] = $variables["form"]["#basic_info"];
+  $variables['basic_info']['#updatelink'] = $variables["form"]["updatelink"];
   $variables["form"]['#attached']['library'][] = 'grants_profile/profile_dialog';
 
   // Unset just in case.
