+++ public/modules/custom/grants_profile/src/Form/GrantsProfileFormRegisteredCommunity.php
@@ -206,18 +206,6 @@ you cannot do any modifications while the form is locked for them.',
 
     $form['#profilecontent'] = $grantsProfileContent;
 
-    $form['updatelink']['link'] = [
-      '#type' => 'submit',
-      '#value' => $this->t('Get updated information', [], $this->tOpts),
-      '#name' => 'refresh_profile',
-      '#submit' => [[$this, 'profileDataRefreshSubmitHandler']],
-      '#ajax' => [
-        'callback' => [$this, 'profileDataRefreshAjaxCallback'],
-        'wrapper' => 'form',
-      ],
-      '#limit_validation_errors' => [],
-    ];
-
     $form_state->setStorage($storage);
 
     $form['actions']['submit_cancel']["#submit"] = [
@@ -225,15 +213,6 @@ you cannot do any modifications while the form is locked for them.',
     ];
 
     return $form;
-  }
-
-  /**
-   * Profile data refresh ajax callback.
-   */
-  public function profileDataRefreshAjaxCallback(array $form) {
-    $response = new AjaxResponse();
-    $response->addCommand(new ReplaceCommand('form', $form));
-    return $response;
   }
 
   /**
