+++ public/modules/custom/grants_profile/templates/grants-profile--basic-info--private-person.html.twig
@@ -29,5 +29,8 @@
     {% if editHelsinkiProfileLink %}
       {{ editHelsinkiProfileLink }} ({{ "The link opens in a new tab"|t({}, {'context': 'grants_profile'}) }})
     {% endif %}
+    <div>
+      {{ updatelink }}
+    </div>
   </div>
 </div>
